Remove old deprecated code.
