+++ raw-compiler/src/main/scala/raw/compiler/base/ProgramContext.scala
@@ -31,8 +31,6 @@ class ProgramContext(
     val compilerContext: CompilerContext
 ) {
 
-  def id: String = runtimeContext.id
-
   def executionLogger: ExecutionLogger = runtimeContext.executionLogger
 
   ///////////////////////////////////////////////////////////////////////////
