+++ raw-compiler/src/main/scala/raw/compiler/base/SyntaxAnalyzer.scala
@@ -16,7 +16,7 @@ import com.typesafe.scalalogging.StrictLogging
 import org.bitbucket.inkytonik.kiama.parsing._
 import org.bitbucket.inkytonik.kiama.util._
 import raw.compiler.base.source._
-import raw.utils.StringEscape
+import raw.utils.RawUtils
 
 import scala.util.matching.Regex
 
@@ -181,12 +181,12 @@ trait SyntaxAnalyzer extends Keywords with StrictLogging {
   final protected lazy val escapedIdent: Parser[String] = escapedIdentRegex ^^ { s => s.drop(1).dropRight(1) }
 
   final protected lazy val stringLitEscaped: Parser[String] = stringLitTripleRegex ^^ { s =>
-    StringEscape.escape(s.drop(3).dropRight(3))
+    RawUtils.escape(s.drop(3).dropRight(3))
   } |
-    stringLitRegex ^^ { s => StringEscape.escape(s.drop(1).dropRight(1)) }
+    stringLitRegex ^^ { s => RawUtils.escape(s.drop(1).dropRight(1)) }
 
   final protected lazy val stringLit: Parser[String] = stringLitTripleRegex ^^ { s => s.drop(3).dropRight(3) } |
-    stringLitRegex ^^ { s => StringEscape.escape(s).drop(1).dropRight(1) }
+    stringLitRegex ^^ { s => RawUtils.escape(s).drop(1).dropRight(1) }
 
   final protected def method1[T](kw: String, p1: => Parser[T]): Parser[T] = s"(?i)$kw\\b".r ~> "(" ~> p1 <~ ")"
