+++ raw-sources-local/src/main/scala/raw/sources/filesystem/local/LocalFileSystem.scala
@@ -14,7 +14,7 @@ package raw.sources.filesystem.local
 
 import raw.sources.bytestream.{DelegatingSeekableInputStream, SeekableInputStream}
 import raw.sources.filesystem._
-import raw.utils.{RawUtils, StringEscape}
+import raw.utils.RawUtils
 
 import java.io._
 import java.nio.file._
@@ -24,7 +24,8 @@ import scala.collection.mutable.ArrayBuffer
 class LocalFileSystem extends BaseFileSystem {
 
   private[sources] val fileSeparator: String = File.separator
-  private val fileSeparatorRegex: String = StringEscape.descape(fileSeparator)
+
+  private val fileSeparatorRegex: String = RawUtils.descape(fileSeparator)
 
   // TODO (msb): This should fail to create if not on developer mode?
   // TODO (msb): And if so, should require all paths to be under some certain base path?
