+++ raw-utils/src/main/resources/reference.conf
@@ -1,13 +1,4 @@
 raw {
   training-wheels = false
   check-syntax-analyzers = false
-
-  jira-crash-reporter {
-    # username = ""
-    # token = ""
-    # project = ""
-    components = ["Compiler"]
-    connection-timeout = 10s
-  }
-
 }
